feat(help): migrate to Typescript
