+++ packages/help/package.json
@@ -2,7 +2,7 @@
   "name": "@axa-fr/react-toolkit-help",
   "version": "1.3.23",
   "main": "dist/index.js",
-  "jsnext:main": "src/index.js",
+  "types": "dist/index.d.ts",
   "private": false,
   "scripts": {
     "prepare": "node ../../scripts/prepare.js"
@@ -13,8 +13,7 @@
   },
   "dependencies": {
     "@axa-fr/react-toolkit-core": "1.3.23",
-    "@axa-fr/react-toolkit-popover": "1.3.23",
-    "prop-types": "15.5.10"
+    "@axa-fr/react-toolkit-popover": "1.3.23"
   },
   "license": "MIT",
   "publishConfig": {
