--- packages/help/src/index.js
@@ -1,3 +1,0 @@
-import Help from './Help.container';
-
-export default Help;
